some fixes

+++ PathFind/Shared.Primitives.Extensions/InclusiveValueRangeExtensions.cs
@@ -16,9 +16,9 @@ namespace Shared.Primitives.Extensions
             }
         }
 
-        public static uint Amplitude(this InclusiveValueRange<int> valueRange)
+        public static long Amplitude(this InclusiveValueRange<int> valueRange)
         {
-            return (uint)((long)valueRange.UpperValueOfRange - valueRange.LowerValueOfRange);
+            return (long)valueRange.UpperValueOfRange - valueRange.LowerValueOfRange;
         }
 
         public static double Amplitude(this InclusiveValueRange<double> valueRange)
